util.Check now writes errors to stderr
Previously, Check threw a fatal error via log for a non-nil error. This
would crash the server any time an error occurs. As an alternative I
wrote CheckHTTP, which writes the error to the HTTP response, but this
would disturb the user experience. Now errors can just be recorded in
the log file to which never writes on the server.